use chan

+++ session/osc.go
@@ -654,41 +654,46 @@ func (s *session) execCommand(r *Record, commandName string, params []string) bo
 
 	buf := bytes.NewBufferString("")
 
-	var stopSign int
-	var lock sync.RWMutex
+	var stdChan = make(chan bool, 2)
 
 	//实时循环读取输出流中的一行内容
 	f := func(reader *bufio.Reader) {
 		for {
 			line, err2 := reader.ReadString('\n')
 			if err2 != nil || io.EOF == err2 {
-				lock.Lock()
-				stopSign++
-				lock.Unlock()
+				stdChan <- true
 				wg.Done()
 				break
 			}
 			buf.WriteString(line)
 			buf.WriteString("\n")
 			s.mysqlAnalyzeOscOutput(line, p)
+
 		}
 	}
-
 	kill := func() {
+		var stopSign int
+		var killOne bool
 		for {
-			if stopSign == 2 {
-				break
-			}
-			if p.Killed {
-				if err := cmd.Process.Kill(); err != nil {
-					s.appendErrorMessage(err.Error())
-				} else {
-					s.appendErrorMessage(fmt.Sprintf("Execute has been abort in percent: %d, remain time: %s",
-						p.Percent, p.RemainTime))
+			select {
+			case <-stdChan:
+				stopSign++
+			default:
+				if stopSign == 2 {
+					goto Loop
 				}
-				break
+				if p.Killed && killOne == false {
+					if err := cmd.Process.Kill(); err != nil {
+						s.appendErrorMessage(err.Error())
+					} else {
+						s.appendErrorMessage(fmt.Sprintf("Execute has been abort in percent: %d, remain time: %s",
+							p.Percent, p.RemainTime))
+					}
+					killOne = true
+				}
 			}
 		}
+	Loop:
 	}
 
 	go kill()
